% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{ar}
\alias{ar}
\title{BlazeAR}
\usage{
ar(y, p = 1, include_mean = TRUE, include_drift = FALSE, xreg = list())
}
\arguments{
\item{y}{The time series to model; an ordered numeric vector or a time-series object.}

\item{p}{The lag order of the model - a non-negative integer.}

\item{include_mean}{Whether to include a mean in the estimation, defaults to TRUE.}

\item{include_drift}{Whether to include drift in the estimation, defaults to FALSE.}

\item{xreg}{A list of exogenous regressors, by default an empty list, please do not set to NULL.}
}
\value{
A fitted model.
}
\description{
Blaze Autoregressive Model Interface
}
\details{
TODO.
}
