% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{arima}
\alias{arima}
\title{BlazeArima}
\usage{
arima(
  y,
  pdq = c(1, 0, 0),
  PDQ = c(0, 0, 0),
  season = 1,
  include_mean = TRUE,
  include_drift = FALSE,
  fitting_method = c("CSS-ML", "CSS", "ML", "LASSO"),
  xreg = list(),
  ss_init = c("Gardner", "Rossignol"),
  transform_parameters = TRUE,
  kappa = 1e+06
)
}
\arguments{
\item{y}{The time series to model; an ordered numeric vector or a time-series object.}

\item{pdq}{The p,d,q order of the model, defaults to c(1,0,0).}

\item{PDQ}{The PDQ (seasonal) order of the model, defaults to c(0,0,0).}

\item{season}{The seasonal period of the time series, defaults to 1}

\item{include_mean}{Whether to include a mean in the estimation, defaults to TRUE.}

\item{include_drift}{Whether to include drift in the estimation, defaults to FALSE.}

\item{fitting_method}{The fitting method to use, one of c("CSS-ML", "CSS", "ML", "LASSO").
See details.}

\item{xreg}{A list of exogenous regressors, by default an empty list, please do not set to NULL.}

\item{ss_init}{The steady state initialization approach, one of c("Gardner", "Rossignol").}

\item{transform_parameters}{Whether to transform parameters when estimating the State Space model,
only impactful for solvers 'CSS-ML' and 'ML'. Defaults to TRUE.}

\item{kappa}{the prior variance (as a multiple of the innovations variance) for the past observations in a differenced model.
Do not reduce this - set to a high value for a reason. Defaults to 1000000.}
}
\value{
A fitted model.
}
\description{
Blaze Arima Model Interface
}
\details{
TODO.
}
