% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ar.R
\name{autoar}
\alias{autoar}
\title{BlazeAutoAR}
\usage{
autoar(
  y,
  min_p = 1,
  max_p = min(length(y) - 1, floor(10 * log10(length(y)))),
  include_mean = TRUE,
  include_drift = FALSE,
  xreg = list(),
  fitting_method = "AIC"
)
}
\arguments{
\item{y}{The time series to model; an ordered numeric vector or a time-series object.}

\item{min_p}{The minimum lag order of the model - a non-negative integer.}

\item{max_p}{The maximum lag order of the model - a non-negative integer.}

\item{include_mean}{Whether to include a mean in the estimation, defaults to TRUE.}

\item{include_drift}{Whether to include drift in the estimation, defaults to FALSE.}

\item{xreg}{A list of exogenous regressors, by default an empty list, please do not set to NULL.}

\item{fitting_method}{The method to use; currently one of "AIC", "AICc", "BIC".}
}
\value{
A fitted model.
}
\description{
Blaze Automatic Autoregressive Model Interface
}
\details{
TODO.
}
