% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaze-package.R
\docType{package}
\name{blaze-package}
\alias{blaze}
\alias{blaze-package}
\title{blaze: Strict Gradual Type Checking}
\description{
A strict, opt-in type system for existing R code. Types are written in ordinary R and enforced at runtime on function arguments, return values, vectors, and data frame columns. The checking engine is implemented in Rust for speed.
}
\author{
\strong{Maintainer}: Brett Gilio \email{giliobrett@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Brett Gilio \email{giliobrett@gmail.com} [copyright holder]
}

}
\keyword{internal}
