% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaze.R
\docType{package}
\name{blaze}
\alias{blaze}
\alias{blaze-package}
\title{Observe and push URL paths}
\usage{
blaze()
}
\description{
The `blaze` package allows a shiny app to simulate the multi-paging
behaviour of a typical web application. Using `blaze` you can walk users
through a shiny app and let them traverse with browsers' forward and back
buttons.

Using path link elements (a variation of standard hyperlinks) users can
browse to different URL paths. A shiny application can detect these changes
with `observePath()` allowing you to update tab sets or other dynamic
elements within the application. `pushPath()` lets you redirect the user from
the server.

Because of how shiny handles URL paths be sure to run the
`paths()` function before launching an application.
}
\section{Including in a shiny app}{


To use blaze with a shiny application the `blaze()` function must be called
inside the application's UI.
}

