% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{blaze_types}
\alias{blaze_types}
\title{List the blaze type vocabulary}
\usage{
blaze_types()
}
\value{
A character vector of the available type, modifier, and refinement
names.
}
\description{
List the blaze type vocabulary
}
\examples{
blaze_types()
}
\seealso{
\code{\link[=type]{type()}}
}
