% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inoise.R
\name{inoise}
\alias{inoise}
\title{BlazeINoise}
\usage{
inoise(y, differences = 1, seasonal_difference = 10, stabilize = TRUE)
}
\arguments{
\item{y}{The time series to model; an ordered numeric vector or a time-series object.}

\item{differences}{The regular differencing order, by default 1.}

\item{seasonal_difference}{The seasonal difference, by default 10.}

\item{stabilize}{Whether to apply a variance stabilizing transform, defaults to TRUE.}
}
\value{
A fitted model.
}
\description{
Blaze Integrated Noise Model Interface
}
\details{
TODO.
}
