% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers.R
\name{observeRoute}
\alias{observeRoute}
\title{Observe routes}
\usage{
observeRoute(path, handler, env = parent.frame(), quoted = FALSE,
  regex = TRUE, domain = getDefaultReactiveDomain())
}
\arguments{
\item{path}{A character string specifying a route path, by default `path` is
treated as a regular expression, see `regex`.}

\item{handler}{An expression or function to call when the uri matches `path`.}

\item{env}{The parent environment of `handler`, defaults to the calling
environment.}

\item{quoted}{One of `TRUE` or `FALSE` specifying if `handler` is a quoted
expression. If `TRUE`, the expression must be quoted with `quote()`.}

\item{regex}{One of `TRUE` or `FALSE` specifying if `path` is treated as a
regular expression, defaults to `TRUE`.}

\item{domain}{A reactive context, defaults to
`shiny::getDefaultReactiveDomain()`.}
}
\description{
Observe a route path and run a handler expression. When the route specified
by `path` is browsed to `handler` will be run.
}
