% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{pathLink}
\alias{pathLink}
\title{Go to paths}
\usage{
pathLink(href, ...)
}
\arguments{
\item{href}{A character string specifying the URL path to navigate to.}

\item{...}{Arguments passed to [htmltools::a()].}
}
\description{
The `pathLink()` function creates a special `tags$a()` element. These links
let the user browse to different parts of your application without refreshing
the current page.
}
