% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{pushPath}
\alias{pushPath}
\alias{getPath}
\title{Path utilities}
\usage{
pushPath(path, session = getDefaultReactiveDomain())

getPath(session = getDefaultReactiveDomain())
}
\arguments{
\item{path}{A character string specifying a new URL path}

\item{session}{A reactive context, defaults to
`shiny::getDefaultReactiveDomain()`.}
}
\description{
Push a new URL path or get the current path.
}
