% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{pushRoute}
\alias{pushRoute}
\alias{getRoute}
\title{Route utilities}
\usage{
pushRoute(path, session = getDefaultReactiveDomain())

getRoute(session = getDefaultReactiveDomain())
}
\arguments{
\item{path}{A character string specifying a new route.}

\item{session}{A reactive context, defaults to
`shiny::getDefaultReactiveDomain()`.}
}
\description{
Push a route or get the current application route.
}
