% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{t_na_ok}
\alias{t_na_ok}
\title{Allow \code{NA} values in a type}
\usage{
t_na_ok(type)
}
\arguments{
\item{type}{A \code{blaze_type}.}
}
\value{
A \code{blaze_type} that allows \code{NA}.
}
\description{
A \code{blaze_type} rejects \code{NA} by default. \code{t_na_ok()} returns a copy that
permits \code{NA}.
}
\examples{
t_na_ok(t_dbl())
}
