% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{t_opt}
\alias{t_opt}
\title{Allow \code{NULL} (an optional value)}
\usage{
t_opt(type)
}
\arguments{
\item{type}{A \code{blaze_type}.}
}
\value{
A \code{blaze_type} that also accepts \code{NULL}.
}
\description{
Returns a copy of \code{type} that also accepts \code{NULL}.
}
\examples{
t_opt(t_int())
}
