% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{base-types}
\alias{base-types}
\alias{t_int}
\alias{t_dbl}
\alias{t_chr}
\alias{t_lgl}
\alias{t_cpl}
\alias{t_raw}
\alias{t_list}
\title{Base type constructors}
\usage{
t_int()

t_dbl()

t_chr()

t_lgl()

t_cpl()

t_raw()

t_list()
}
\value{
A \code{blaze_type} object.
}
\description{
Each constructor returns a \code{blaze_type} describing an R vector of a single
base type, named as \code{\link[=typeof]{typeof()}} reports it.
}
\examples{
t_int()
t_chr()
}
