% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blaze.R
\docType{package}
\name{blaze-package}
\alias{blaze}
\alias{blaze-package}
\title{Route paths and URIs}
\description{
The `blaze` package makes simulating different routes within a shiny
application possible. The approach makes use of the `#`, hash, portition of
the uri to prevent the application from truly reloading.
}
