% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{pathLink}
\alias{pathLink}
\title{Go to paths}
\usage{
pathLink(..., href)
}
\description{
The `pathLink()` function creates a special `tags$a()` element. These links
let the user browse to different parts of your application without refreshing
the current page.
}
