% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{pushPath}
\alias{pushPath}
\alias{getPath}
\title{Path utilities}
\usage{
pushPath(
  path,
  mode = c("push", "replace"),
  session = getDefaultReactiveDomain()
)

getPath(session = getDefaultReactiveDomain())
}
\arguments{
\item{path}{A character string specifying a new URL path}

\item{mode}{Either `"push"` or `"replace"`. If `"push"`, the default, the
path is pushed onto the history stack and pressing the back button in the
browser will redirect to the current path (before pushing the path). If
`"replace"`, then the pushed path will replace the current path without
changing the next page in the browser's back button stack.}

\item{session}{A reactive context, defaults to
`shiny::getDefaultReactiveDomain()`.}
}
\description{
Push a new URL path or get the current path.
}
